# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/person": {
      "get": {
        "operationId": "getAllPeople",
        "parameters": [
          {
            "name": "if-none-match",
            "in": "header",
            "description": "ETag of data already cached on the client",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withHidden",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      },
      "post": {
        "operationId": "createPerson",
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      },
      "put": {
        "operationId": "updatePeople",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeopleUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BulkIdResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}": {
      "get": {
        "operationId": "getPerson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      },
      "put": {
        "operationId": "updatePerson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}/assets": {
      "get": {
        "operationId": "getPersonAssets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}/merge": {
      "post": {
        "operationId": "mergePerson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergePersonDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BulkIdResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}/reassign": {
      "put": {
        "operationId": "reassignFaces",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetFaceUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}/statistics": {
      "get": {
        "operationId": "getPersonStatistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonStatisticsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
    "/person/{id}/thumbnail": {
      "get": {
        "operationId": "getPersonThumbnail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Person"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.94.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AssetFaceUpdateDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceUpdateItem"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "AssetFaceUpdateItem": {
        "properties": {
          "assetId": {
            "format": "uuid",
            "type": "string"
          },
          "personId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "assetId",
          "personId"
        ],
        "type": "object"
      },
      "AssetFaceWithoutPersonResponseDto": {
        "properties": {
          "boundingBoxX1": {
            "type": "integer"
          },
          "boundingBoxX2": {
            "type": "integer"
          },
          "boundingBoxY1": {
            "type": "integer"
          },
          "boundingBoxY2": {
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageHeight": {
            "type": "integer"
          },
          "imageWidth": {
            "type": "integer"
          }
        },
        "required": [
          "boundingBoxX1",
          "boundingBoxX2",
          "boundingBoxY1",
          "boundingBoxY2",
          "id",
          "imageHeight",
          "imageWidth"
        ],
        "type": "object"
      },
...
      "AssetResponseDto": {
        "properties": {
          "checksum": {
            "description": "base64 encoded sha1 hash",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "hasMetadata": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isExternal": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isReadOnly": {
            "type": "boolean"
          },
          "isTrashed": {
            "type": "boolean"
          },
          "libraryId": {
            "type": "string"
          },
          "livePhotoVideoId": {
            "nullable": true,
            "type": "string"
          },
          "localDateTime": {
            "format": "date-time",
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonWithFacesResponseDto"
            },
            "type": "array"
          },
          "resized": {
            "type": "boolean"
          },
          "smartInfo": {
            "$ref": "#/components/schemas/SmartInfoResponseDto"
          },
          "stack": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "stackCount": {
            "nullable": true,
            "type": "integer"
          },
          "stackParentId": {
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            },
            "type": "array"
          },
          "thumbhash": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AssetTypeEnum"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "deviceAssetId",
          "deviceId",
          "duration",
          "fileCreatedAt",
          "fileModifiedAt",
          "hasMetadata",
          "id",
          "isArchived",
          "isExternal",
          "isFavorite",
          "isOffline",
          "isReadOnly",
          "isTrashed",
          "libraryId",
          "localDateTime",
          "originalFileName",
          "originalPath",
          "ownerId",
          "resized",
          "stackCount",
          "thumbhash",
          "type",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AssetTypeEnum": {
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ],
        "type": "string"
      },
...
      "BulkIdResponseDto": {
        "properties": {
          "error": {
            "enum": [
              "duplicate",
              "no_permission",
              "not_found",
              "unknown"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "success"
        ],
        "type": "object"
      },
...
      "ExifResponseDto": {
        "properties": {
          "city": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "country": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "dateTimeOriginal": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "exifImageHeight": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exifImageWidth": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exposureTime": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "fNumber": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "fileSizeInByte": {
            "default": null,
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "focalLength": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "iso": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "latitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "lensModel": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "longitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "make": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "model": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "modifyDate": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "orientation": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "projectionType": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "state": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "timeZone": {
            "default": null,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MergePersonDto": {
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
...
      "PeopleResponseDto": {
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonResponseDto"
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "people",
          "total"
        ],
        "type": "object"
      },
      "PeopleUpdateDto": {
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/PeopleUpdateItem"
            },
            "type": "array"
          }
        },
        "required": [
          "people"
        ],
        "type": "object"
      },
      "PeopleUpdateItem": {
        "properties": {
          "birthDate": {
            "description": "Person date of birth.\nNote: the mobile app cannot currently set the birth date to null.",
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "featureFaceAssetId": {
            "description": "Asset is used to get the feature face thumbnail.",
            "type": "string"
          },
          "id": {
            "description": "Person id.",
            "type": "string"
          },
          "isHidden": {
            "description": "Person visibility",
            "type": "boolean"
          },
          "name": {
            "description": "Person name.",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "PersonResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "id",
          "isHidden",
          "name",
          "thumbnailPath"
        ],
        "type": "object"
      },
      "PersonStatisticsResponseDto": {
        "properties": {
          "assets": {
            "type": "integer"
          }
        },
        "required": [
          "assets"
        ],
        "type": "object"
      },
      "PersonUpdateDto": {
        "properties": {
          "birthDate": {
            "description": "Person date of birth.\nNote: the mobile app cannot currently set the birth date to null.",
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "featureFaceAssetId": {
            "description": "Asset is used to get the feature face thumbnail.",
            "type": "string"
          },
          "isHidden": {
            "description": "Person visibility",
            "type": "boolean"
          },
          "name": {
            "description": "Person name.",
            "type": "string"
          }
        },
        "type": "object"
      },
      "PersonWithFacesResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "faces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "faces",
          "id",
          "isHidden",
          "name",
          "thumbnailPath"
        ],
        "type": "object"
      },
...
      "SmartInfoResponseDto": {
        "properties": {
          "objects": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "TagResponseDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TagTypeEnum"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "userId"
        ],
        "type": "object"
      },
      "TagTypeEnum": {
        "enum": [
          "OBJECT",
          "FACE",
          "CUSTOM"
        ],
        "type": "string"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserResponseDto": {
        "properties": {
          "avatarColor": {
            "$ref": "#/components/schemas/UserAvatarColor"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "externalPath": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "memoriesEnabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "oauthId": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "quotaSizeInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "quotaUsageInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "storageLabel": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "createdAt",
          "deletedAt",
          "email",
          "externalPath",
          "id",
          "isAdmin",
          "name",
          "oauthId",
          "profileImagePath",
          "quotaSizeInBytes",
          "quotaUsageInBytes",
          "shouldChangePassword",
          "storageLabel",
          "updatedAt"
        ],
        "type": "object"
      },
...
    }
  }
}